--- addon/components/ember-table-body.js
@@ -1,32 +1,0 @@
-import Component from '@ember/component';
-import layout from '../templates/components/ember-table-body';
-
-import { tagName } from '@ember-decorators/component';
-import { argument } from '@ember-decorators/argument';
-import { required } from '@ember-decorators/argument/validation';
-import { type } from '@ember-decorators/argument/type';
-
-@tagName('')
-export default class EmberTableBody extends Component {
-  layout = layout;
-
-  @argument
-  @required
-  @type('object')
-  api;
-
-  @argument
-  @required
-  @type('object')
-  rows;
-
-  @argument
-  @required
-  @type('number')
-  estimateRowHeight;
-
-  @argument
-  @required
-  @type('boolean')
-  staticHeight;
-}
