+++ addon/components/ember-table-cell.js
@@ -4,6 +4,7 @@ import { argument } from '@ember-decorators/argument';
 import { required } from '@ember-decorators/argument/validation';
 import { type } from '@ember-decorators/argument/type';
 import { Action } from '@ember-decorators/argument/types';
+import { get } from '@ember/object';
 
 import layout from '../templates/components/ember-table-cell';
 
@@ -26,4 +27,13 @@ export default class EmberTableCell extends EmberTableBaseCell {
   @type(Action)
   onChecked;
 
+  click() {
+    let rowValue = this.get('cell.row.value');
+    let rowIndex = this.get('cell.row.index');
+    let api = this.get('cell.row.api');
+
+    if (this.get('columnIndex') === 0 && Array.isArray(get(rowValue, 'children'))) {
+      api.toggleRowCollapse(rowIndex);
+    }
+  }
 }
