--- addon/components/tree-table-body.js
@@ -1,6 +1,0 @@
-import EmberTableBody from './ember-table-body';
-import layout from '../templates/components/tree-table-body';
-
-export default class TreeTableBody extends EmberTableBody {
-  layout = layout;
-}
