--- addon/components/tree-table-row.js
@@ -1,12 +1,0 @@
-import EmberTableRow from './ember-table-row';
-
-export default class TreeTableRow extends EmberTableRow {
-  /**
-   * @override
-   */
-  _outerCellComponent = 'tree-table-cell';
-
-  toggleRowCollapse(index) {
-    this.get('api').toggleRowCollapse(index);
-  }
-}
