--- addon/templates/components/ember-table-body.hbs
@@ -1,13 +1,0 @@
-{{#vertical-collection rows
-  estimateHeight=estimateRowHeight
-  staticHeight=staticHeight
-  containerSelector=".ember-table"
-
-  as |rowValue rowIndex|
-}}
-  {{yield (hash
-    value=rowValue
-    index=rowIndex
-    api=api)
-  }}
-{{/vertical-collection}}
