--- addon/templates/components/tree-table-body.hbs
@@ -1,14 +1,0 @@
-{{#vertical-collection rows
-  estimateHeight=estimateRowHeight
-  staticHeight=staticHeight
-  containerSelector=".ember-table"
-
-  as |rowValue rowIndex|
-}}
-  {{yield (hash
-    value=rowValue.value
-    depth=rowValue.depth
-    index=rowIndex
-    api=api)
-  }}
-{{/vertical-collection}}
