--- app/components/tree-table-body.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-table/components/tree-table-body';
