--- app/components/tree-table-cell.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-table/components/tree-table-cell';
