--- app/components/tree-table-row.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-table/components/tree-table-row';
